% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{prepare_ncrna_rpf}
\alias{prepare_ncrna_rpf}
\title{Prepare ncRNA ribosome footprints}
\usage{
prepare_ncrna_rpf(lst_rpf, tx_info, par_learn)
}
\arguments{
\item{lst_rpf}{Output of \code{\link[=prep_rpf]{prep_rpf()}} for ncRNA alignments.}

\item{tx_info}{Transcript annotation list for ncRNAs.}

\item{par_learn}{Parameter list returned by \code{\link[=learn_bias_parameters]{learn_bias_parameters()}} based on CDS data.}
}
\value{
A list containing processed footprints and inherited parameters.
}
\description{
Applies the learnt parameters from CDS libraries to ncRNA-aligned footprints,
returning a harmonised object that can be passed to \code{\link[=predict_orf]{predict_orf()}}.
}
