% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RiboBA-package}
\alias{RiboBA}
\alias{RiboBA-package}
\title{RiboBA: Bias-Aware Framework for Robust ORF Detection}
\description{
RiboBA is a bias-aware framework for robust ORF detection across diverse ribosome profiling datasets.
}
\author{
\strong{Maintainer}: Junyu Bai \email{bai.junyu.bio@gmail.com}

}
\keyword{internal}
